feat(api): Enhance pharmacist controllers with new features

+++ src/main/java/com/his/dto/PharmacistStatisticsDTO.java
@@ -11,7 +11,7 @@ public class PharmacistStatisticsDTO {
     /**
      * 今日发药单数
      */
-    private Integer dispensedCount;
+    private Long dispensedCount;
 
     /**
      * 今日发药总金额
@@ -21,17 +21,17 @@ public class PharmacistStatisticsDTO {
     /**
      * 今日发药药品总数
      */
-    private Integer totalItems;
+    private Long totalItems;
 
     public PharmacistStatisticsDTO() {
-        this.dispensedCount = 0;
+        this.dispensedCount = 0L;
         this.totalAmount = BigDecimal.ZERO;
-        this.totalItems = 0;
+        this.totalItems = 0L;
     }
     
-    public PharmacistStatisticsDTO(Integer dispensedCount, BigDecimal totalAmount, Integer totalItems) {
-        this.dispensedCount = dispensedCount != null ? dispensedCount : 0;
+    public PharmacistStatisticsDTO(Long dispensedCount, BigDecimal totalAmount, Long totalItems) {
+        this.dispensedCount = dispensedCount != null ? dispensedCount : 0L;
         this.totalAmount = totalAmount != null ? totalAmount : BigDecimal.ZERO;
-        this.totalItems = totalItems != null ? totalItems : 0;
+        this.totalItems = totalItems != null ? totalItems : 0L;
     }
 }
